Extract mock_trade generation to sepearate file
